Add files via upload
Process-informed Nonstationary Extreme Value Analysis (ProNEVA) is a Matlab software package designed to facilitate extreme value analysis (EVA) under both stationary and nonstationary assumptions. ProNEVA allows using time or a physically-based covariate to describe changes in statistics of extremes. Examples of physically-based covariates include: change in runoff extremes in response to urbanization, or change in temperature extremes as a function of CO2 in the atmosphere. ProNEVA estimates the parameters of the Generalized Extreme Value (GEV), the Generalized Pareto (GP), and the Log-Pearson Type III (LP3) distributions. The model includes a Bayesian approach and a hybrid Markov Chain Monte Carlo (MCMC) method for sampling from the posterior distribution. ProNEVA also provides diagnostic tests and return period-return level plots. The source code of the toolbox is released along with a Graphical User Interface (GUI). By using ProNEVA, the users agree to the disclaimer included in the package.